# src/main/java/org/example/cuidadodemascotas/usermicroservice/apis/controller/UserApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.example.cuidadodemascotas.usermicroservice.apis.controller;

import org.example.cuidadodemascotas.usermicroservice.apis.dto.ApiError;
import org.example.cuidadodemascotas.usermicroservice.apis.dto.UserPageResponse;
import org.example.cuidadodemascotas.usermicroservice.apis.dto.UserRequestDTO;
import org.example.cuidadodemascotas.usermicroservice.apis.dto.UserResponseDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-10-27T13:15:11.948410900-03:00[America/Asuncion]")
@Validated
@Tag(name = "User", description = "the User API")
public interface UserApi {

    /**
     * GET /users/count : Contar todos los usuarios activos
     *
     * @return Conteo de usuarios (status code 200)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "countUsers",
        summary = "Contar todos los usuarios activos",
        tags = { "User" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Conteo de usuarios", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Long.class))
            }),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/count",
        produces = { "application/json" }
    )
    
    ResponseEntity<Long> countUsers(
        
    );


    /**
     * GET /users/count/role/{roleId} : Contar usuarios por role
     *
     * @param roleId  (required)
     * @return Conteo de usuarios (status code 200)
     *         or Role no encontrado (status code 404)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "countUsersByRole",
        summary = "Contar usuarios por role",
        tags = { "User" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Conteo de usuarios", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Long.class))
            }),
            @ApiResponse(responseCode = "404", description = "Role no encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            }),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/count/role/{roleId}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Long> countUsersByRole(
        @Parameter(name = "roleId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("roleId") Long roleId
    );


    /**
     * POST /users : Crear un nuevo usuario
     *
     * @param userRequestDTO  (required)
     * @return Usuario creado (status code 201)
     *         or Solicitud inválida (status code 400)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "createUser",
        summary = "Crear un nuevo usuario",
        tags = { "User" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Usuario creado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserResponseDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Solicitud inválida", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            }),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/users",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<UserResponseDTO> createUser(
        @Parameter(name = "UserRequestDTO", description = "", required = true) @Valid @RequestBody UserRequestDTO userRequestDTO
    );


    /**
     * DELETE /users/{id} : Eliminar usuario (soft delete)
     *
     * @param id  (required)
     * @return Usuario eliminado (status code 204)
     *         or Usuario no encontrado (status code 404)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "deleteUser",
        summary = "Eliminar usuario (soft delete)",
        tags = { "User" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Usuario eliminado"),
            @ApiResponse(responseCode = "404", description = "Usuario no encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            }),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/users/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Void> deleteUser(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * GET /users/carers : Obtener todos los cuidadores
     *
     * @return Cuidadores obtenidos (status code 200)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "getCarers",
        summary = "Obtener todos los cuidadores",
        tags = { "User" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Cuidadores obtenidos", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = UserResponseDTO.class)))
            }),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/carers",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<UserResponseDTO>> getCarers(
        
    );


    /**
     * GET /users/carers/paged : Obtener cuidadores con paginación
     *
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 20)
     * @return Cuidadores obtenidos (status code 200)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "getCarersPaged",
        summary = "Obtener cuidadores con paginación",
        tags = { "User" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Cuidadores obtenidos", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserPageResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/carers/paged",
        produces = { "application/json" }
    )
    
    ResponseEntity<UserPageResponse> getCarersPaged(
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size
    );


    /**
     * GET /users/owners : Obtener todos los dueños
     *
     * @return Dueños obtenidos (status code 200)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "getOwners",
        summary = "Obtener todos los dueños",
        tags = { "User" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Dueños obtenidos", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = UserResponseDTO.class)))
            }),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/owners",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<UserResponseDTO>> getOwners(
        
    );


    /**
     * GET /users/owners/paged : Obtener dueños con paginación
     *
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 20)
     * @return Dueños obtenidos (status code 200)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "getOwnersPaged",
        summary = "Obtener dueños con paginación",
        tags = { "User" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Dueños obtenidos", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserPageResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/owners/paged",
        produces = { "application/json" }
    )
    
    ResponseEntity<UserPageResponse> getOwnersPaged(
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size
    );


    /**
     * GET /users/email/{email} : Obtener usuario por email
     *
     * @param email  (required)
     * @return Usuario obtenido (status code 200)
     *         or Usuario no encontrado (status code 404)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "getUserByEmail",
        summary = "Obtener usuario por email",
        tags = { "User" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Usuario obtenido", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserResponseDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "Usuario no encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            }),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/email/{email}",
        produces = { "application/json" }
    )
    
    ResponseEntity<UserResponseDTO> getUserByEmail(
        @Parameter(name = "email", description = "", required = true, in = ParameterIn.PATH) @PathVariable("email") String email
    );


    /**
     * GET /users/{id} : Obtener usuario por ID
     *
     * @param id  (required)
     * @return Usuario obtenido (status code 200)
     *         or Usuario no encontrado (status code 404)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "getUserById",
        summary = "Obtener usuario por ID",
        tags = { "User" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Usuario obtenido", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserResponseDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "Usuario no encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            }),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<UserResponseDTO> getUserById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * GET /users : Obtener todos los usuarios con paginación
     *
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 20)
     * @return Usuarios obtenidos (status code 200)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "getUsers",
        summary = "Obtener todos los usuarios con paginación",
        tags = { "User" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Usuarios obtenidos", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserPageResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users",
        produces = { "application/json" }
    )
    
    ResponseEntity<UserPageResponse> getUsers(
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size
    );


    /**
     * GET /users/role/{roleId} : Obtener usuarios por role con paginación
     *
     * @param roleId  (required)
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 20)
     * @return Usuarios obtenidos (status code 200)
     *         or Role no encontrado (status code 404)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "getUsersByRole",
        summary = "Obtener usuarios por role con paginación",
        tags = { "User" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Usuarios obtenidos", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserPageResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Role no encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            }),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/role/{roleId}",
        produces = { "application/json" }
    )
    
    ResponseEntity<UserPageResponse> getUsersByRole(
        @Parameter(name = "roleId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("roleId") Long roleId,
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size
    );


    /**
     * GET /users/search/{text} : Buscar usuarios por nombre o apellido
     *
     * @param text  (required)
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 10)
     * @return Usuarios encontrados (status code 200)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "searchUsers",
        summary = "Buscar usuarios por nombre o apellido",
        tags = { "User" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Usuarios encontrados", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserPageResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/search/{text}",
        produces = { "application/json" }
    )
    
    ResponseEntity<UserPageResponse> searchUsers(
        @Parameter(name = "text", description = "", required = true, in = ParameterIn.PATH) @PathVariable("text") String text,
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "10") Integer size
    );


    /**
     * PUT /users/{id} : Actualizar usuario
     *
     * @param id  (required)
     * @param userRequestDTO  (required)
     * @return Usuario actualizado (status code 200)
     *         or Solicitud inválida (status code 400)
     *         or Usuario no encontrado (status code 404)
     *         or Error interno del servidor (status code 500)
     */
    @Operation(
        operationId = "updateUser",
        summary = "Actualizar usuario",
        tags = { "User" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Usuario actualizado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserResponseDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Solicitud inválida", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            }),
            @ApiResponse(responseCode = "404", description = "Usuario no encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            }),
            @ApiResponse(responseCode = "500", description = "Error interno del servidor", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/users/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<UserResponseDTO> updateUser(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "UserRequestDTO", description = "", required = true) @Valid @RequestBody UserRequestDTO userRequestDTO
    );

}